requirements.txt
# VaultApp V5 依赖环境锁定 (详见设计文档第1节)
kivy==2.3.0
kivymd==1.2.0
cryptography==42.0.8
webdav4==0.9.8
pyjnius==1.6.1
httpx==0.27.0
